fix: handshake can be bypassed from the ui
