+++ src/OneLiner.tsx
@@ -4,18 +4,16 @@
 import { FC, useCallback, useEffect, useState } from "react";
 import { KaspaClient } from "./utils/all-in-one";
 import { unknownErrorToErrorLike } from "./utils/errors";
-import { Contact, Message, NetworkType } from "./types/all";
+import { Contact, NetworkType } from "./types/all";
 import { useMessagingStore } from "./store/messaging.store";
 import { ContactCard } from "./components/ContactCard";
-import { MessageDisplay } from "./components/MessageDisplay";
 import { WalletInfo } from "./components/WalletInfo";
-import { SendMessageForm } from "./containers/SendMessageForm";
 import { ErrorCard } from "./components/ErrorCard";
 import { useWalletStore } from "./store/wallet.store";
 import { WalletGuard } from "./containers/WalletGuard";
-import { FetchApiMessages } from "./components/FetchApiMessages";
 import { NewChatForm } from "./components/NewChatForm";
 import styles from "./OneLiner.module.css";
+import { MessageSection } from "./containers/MessagesSection";
 
 export const OneLiner: FC = () => {
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
@@ -210,79 +208,6 @@ export const OneLiner: FC = () => {
     }
   }, [walletStore.unlockedWallet, messageStore]);
 
-  const onClearHistory = useCallback(() => {
-    if (!walletStore.address) {
-      return;
-    }
-
-    if (
-      confirm(
-        "Are you sure you want to clear all message history? This cannot be undone."
-      )
-    ) {
-      messageStore.flushCache(walletStore.address.toString());
-    }
-  }, [walletStore.address, messageStore]);
-
-  const onExportMessages = useCallback(async () => {
-    if (!walletStore.unlockedWallet?.password) {
-      alert("Please unlock your wallet first");
-      return;
-    }
-
-    try {
-      const blob = await messageStore.exportMessages(
-        walletStore.unlockedWallet,
-        walletStore.unlockedWallet.password
-      );
-
-      // Create download link
-      const url = URL.createObjectURL(blob);
-      const a = document.createElement("a");
-      a.href = url;
-      a.download = `kaspa-messages-backup-${
-        new Date().toISOString().split("T")[0]
-      }.json`;
-      document.body.appendChild(a);
-      a.click();
-      document.body.removeChild(a);
-      URL.revokeObjectURL(url);
-    } catch (error) {
-      console.error("Error exporting messages:", error);
-      alert("Failed to export messages");
-    }
-  }, [messageStore, walletStore.unlockedWallet]);
-
-  const onImportMessages = useCallback(
-    async (event: React.ChangeEvent<HTMLInputElement>) => {
-      const file = event.target.files?.[0];
-      if (!file) return;
-
-      if (!walletStore.unlockedWallet?.password) {
-        alert("Please unlock your wallet first");
-        return;
-      }
-
-      try {
-        await messageStore.importMessages(
-          file,
-          walletStore.unlockedWallet,
-          walletStore.unlockedWallet.password
-        );
-        alert("Messages imported successfully!");
-      } catch (error: unknown) {
-        console.error("Error importing messages:", error);
-        alert(
-          error instanceof Error ? error.message : "Failed to import messages"
-        );
-      }
-
-      // Clear the input
-      event.target.value = "";
-    },
-    [messageStore, walletStore.unlockedWallet]
-  );
-
   const onStartMessagingProcessClicked = useCallback(async () => {
     try {
       if (!currentClient || !currentClient.connected) {
@@ -325,34 +250,6 @@ export const OneLiner: FC = () => {
       messageStore.setOpenedRecipient(contact.address);
     },
     [messageStore, walletStore.address]
-  );
-
-  // Add function to handle handshake initiation
-  const initiateHandshakeWithRecipient = useCallback(
-    async (recipientAddress: string) => {
-      try {
-        if (!walletStore.unlockedWallet?.password) {
-          throw new Error("Wallet must be unlocked");
-        }
-
-        // Initiate handshake and get payload
-        const { payload, conversation } = await messageStore.initiateHandshake(
-          recipientAddress
-        );
-
-        // Create and send transaction with handshake payload
-        // You'll need to implement the actual transaction sending here
-        // This might involve using your existing transaction creation methods
-
-        messageStore.setIsCreatingNewChat(false);
-      } catch (error) {
-        console.error("Failed to initiate handshake:", error);
-        setErrorMessage(
-          `Failed to initiate handshake: ${unknownErrorToErrorLike(error)}`
-        );
-      }
-    },
-    [walletStore.unlockedWallet, messageStore]
   );
 
   return (
@@ -434,69 +331,7 @@ export const OneLiner: FC = () => {
                 ))}
             </div>
           </div>
-          <div className="messages-section">
-            <div className="messages-header">
-              <h3>Messages</h3>
-              <div className="header-actions">
-                {walletStore.address && (
-                  <FetchApiMessages address={walletStore.address.toString()} />
-                )}
-                <button
-                  onClick={onExportMessages}
-                  className="backup-button"
-                  title="Export message backup"
-                >
-                  Export Backup
-                </button>
-                <label className="import-button" title="Import message backup">
-                  Import Backup
-                  <input
-                    type="file"
-                    accept=".json"
-                    style={{ display: "none" }}
-                    onChange={onImportMessages}
-                  />
-                </label>
-                <button
-                  onClick={onClearHistory}
-                  id="clearHistoryButton"
-                  className="clear-history-button"
-                >
-                  Clear History
-                </button>
-              </div>
-            </div>
-            <div
-              className="messages-list"
-              ref={(el) => {
-                // Auto-scroll to bottom when new messages arrive
-                if (el) {
-                  el.scrollTop = el.scrollHeight;
-                }
-              }}
-            >
-              {messageStore.isCreatingNewChat ? (
-                <div className="no-messages">
-                  Enter a recipient address to start a new conversation.
-                </div>
-              ) : messageStore.messagesOnOpenedRecipient.length ? (
-                messageStore.messagesOnOpenedRecipient.map((msg) => (
-                  <MessageDisplay
-                    isOutgoing={
-                      msg.senderAddress === walletStore.address?.toString()
-                    }
-                    key={msg.transactionId}
-                    message={msg}
-                  />
-                ))
-              ) : (
-                <div className="no-messages">
-                  No messages in this conversation.
-                </div>
-              )}
-            </div>
-            <SendMessageForm />
-          </div>
+          <MessageSection />
         </div>
       ) : null}
       <div id="transactions">
